Deformable-3D-Gaussians/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
absl-py=2.1.0=pypi_0
accelerate=0.32.1=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
blas=1.1=openblas
brotli-python=1.0.9=py312h6a678d5_8
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.7.4=hbcca054_0
certifi=2024.7.4=pyhd8ed1ab_0
cffi=1.16.0=py312h5eee18b_1
charset-normalizer=3.3.2=pyhd8ed1ab_0
chex=0.1.86=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.2.2=pyhd8ed1ab_0
contourpy=1.2.1=pypi_0
cycler=0.12.1=pypi_0
dearpygui=1.11.1=pypi_0
debugpy=1.6.7=py312h6a678d5_0
decorator=5.1.1=pyhd8ed1ab_0
diff-gaussian-rasterization=0.0.0=pypi_0
diffusers=0.29.2=pyhd8ed1ab_0
etils=1.9.2=pypi_0
exceptiongroup=1.2.2=pyhd8ed1ab_0
executing=2.0.1=pyhd8ed1ab_0
expat=2.6.2=h6a678d5_0
filelock=3.15.4=pyhd8ed1ab_0
flax=0.8.5=pypi_0
fonttools=4.53.1=pypi_0
freetype=2.10.4=h0708190_1
fsspec=2024.6.1=pyhff2d567_0
h2=4.1.0=pyhd8ed1ab_0
hpack=4.0.0=pyh9f0ad1d_0
huggingface_hub=0.23.4=pyhd8ed1ab_0
hyperframe=6.0.1=pyhd8ed1ab_0
idna=3.7=pyhd8ed1ab_0
imageio=2.27.0=pypi_0
imageio-ffmpeg=0.5.1=pypi_0
importlib-metadata=8.0.0=pyha770c72_0
importlib-resources=6.4.0=pypi_0
importlib_metadata=8.0.0=hd8ed1ab_0
ipykernel=6.29.5=pyh3099207_0
ipython=8.26.0=pyh707e725_0
jax=0.4.30=pypi_0
jaxlib=0.4.30=pypi_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.4=pyhd8ed1ab_0
jpeg=9e=h0b41bf4_3
jupyter_client=8.6.2=pyhd8ed1ab_0
jupyter_core=5.7.2=py312h06a4308_0
kiwisolver=1.4.5=pypi_0
lcms2=2.15=hfd0df8a_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h9c3ff4c_0
libdeflate=1.17=h0b41bf4_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=14.1.0=h77fa898_0
libgfortran-ng=14.1.0=h69a702a_0
libgfortran5=14.1.0=hc5f4f2c_0
libgomp=14.1.0=h77fa898_0
libopenblas=0.3.27=pthreads_hac2b453_1
libpng=1.6.39=h5eee18b_0
libprotobuf=3.20.3=he621ea3_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.5.1=h6a678d5_0
libuuid=1.41.5=h5eee18b_0
libwebp-base=1.4.0=hd590300_0
lpips=0.1.4=pypi_0
lz4-c=1.9.4=h6a678d5_1
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=py312h5eee18b_0
matplotlib=3.9.1=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mdurl=0.1.2=pypi_0
ml-dtypes=0.4.0=pypi_0
mpmath=1.3.0=pyhd8ed1ab_0
msgpack=1.0.8=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.3=pyhd8ed1ab_1
numpy=1.26.4=py312h2809609_0
numpy-base=1.26.4=py312he1a6c75_0
nvidia-cublas-cu11=11.11.3.6=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu11=11.8.87=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu11=11.8.89=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu11=11.8.89=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu11=8.7.0.84=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu11=10.3.0.86=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu11=11.4.1.48=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu11=11.7.5.86=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu11=2.20.5=pypi_0
nvidia-nccl-cu12=2.20.5=pypi_0
nvidia-nvjitlink-cu12=12.5.82=pypi_0
nvidia-nvtx-cu11=11.8.86=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
openblas=0.3.27=pthreads_h9eca1d5_1
opencv-python=4.10.0.84=pypi_0
openjpeg=2.4.0=h9ca470c_1
openssl=3.3.1=h4ab18f5_1
opt-einsum=3.3.0=pypi_0
optax=0.2.3=pypi_0
orbax-checkpoint=0.5.20=pypi_0
packaging=24.1=pyhd8ed1ab_0
parso=0.8.4=pyhd8ed1ab_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.3.0=py312h5eee18b_0
pip=24.0=py312h06a4308_0
platformdirs=4.2.2=pyhd8ed1ab_0
plyfile=0.8.1=pypi_0
prompt-toolkit=3.0.47=pyha770c72_0
protobuf=5.27.2=pypi_0
psutil=6.0.0=pypi_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pycparser=2.22=pyhd8ed1ab_0
pygments=2.18.0=pyhd8ed1ab_0
pyparsing=3.1.2=pypi_0
pysocks=1.7.1=pyha2e5f31_6
python=3.12.4=h5148396_1
python-dateutil=2.9.0=pyhd8ed1ab_0
pyyaml=6.0.1=py312h5eee18b_0
pyzmq=25.1.2=py312h6a678d5_0
readline=8.2=h5eee18b_0
regex=2024.5.15=pypi_0
requests=2.32.3=pyhd8ed1ab_0
rich=13.7.1=pypi_0
safetensors=0.4.3=pypi_0
scipy=1.14.0=pypi_0
sentencepiece=0.2.0=pypi_0
setuptools=69.5.1=py312h06a4308_0
simple-knn=0.0.0=pypi_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.45.3=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.12.1=pyh04b8f61_3
tensorstore=0.1.63=pypi_0
timm=1.0.7=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.19.1=pypi_0
toolz=0.12.1=pypi_0
torch=2.3.1+cu118=pypi_0
torchaudio=2.3.1+cu118=pypi_0
torchvision=0.18.1+cu118=pypi_0
tornado=6.4.1=py312h5eee18b_0
tqdm=4.66.4=pyhd8ed1ab_0
traitlets=5.14.3=pyhd8ed1ab_0
transformers=4.43.0.dev0=pypi_0
typing-extensions=4.12.2=hd8ed1ab_0
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024a=h04d1e81_0
urllib3=2.2.2=pyhd8ed1ab_1
wcwidth=0.2.13=pyhd8ed1ab_0
wheel=0.43.0=py312h06a4308_0
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7f98852_2
zeromq=4.3.5=h6a678d5_0
zipp=3.19.2=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_1
zstandard=0.22.0=py312h2c38b39_0
zstd=1.5.5=hc292b87_2
